requirements.txt
langchain
langchain-openai
langchain-community
python-dotenv
streamlit